Fix Put Shelly in separat file #13
