+++ include/GlobalDefinitions.h
@@ -67,8 +67,6 @@ typedef struct {
 } BrunnenData;
 
 
-void switchPowerOff();
-
 
 /*
 // JSON-Document für WebService und Kommunikation mit html-page und mqtt
