+++ src/JSONMessage.cpp
@@ -38,6 +38,7 @@
 #include "KSWS2812B.h"
 #include "Fontain.h"
 #include "Settings.h"
+#include "Shelly.h"		// we controll the power of the LEDs and Fontain with a shelly at 220V side
 
 
 extern float temperature;
